Fix a defect reading for multiple extents.

+++ fs/src/fs/org/jnode/fs/hfsplus/HfsPlusForkData.java
@@ -33,11 +33,11 @@ public class HfsPlusForkData {
      */
     private long totalSize;
     /** */
-    private long clumpSize;
+    private int clumpSize;
     /**
      * The total of allocation blocks use by the extents in the fork.
      */
-    private long totalBlock;
+    private int totalBlock;
     /**
      * The first eight extent descriptors for the fork.
      */
@@ -53,8 +53,8 @@ public class HfsPlusForkData {
         byte[] data = new byte[FORK_DATA_LENGTH];
         System.arraycopy(src, offset, data, 0, FORK_DATA_LENGTH);
         totalSize = BigEndian.getInt64(data, 0);
-        clumpSize = BigEndian.getUInt32(data, 8);
-        totalBlock = BigEndian.getUInt32(data, 12);
+        clumpSize = BigEndian.getInt32(data, 8);
+        totalBlock = BigEndian.getInt32(data, 12);
         extents = new ExtentDescriptor[8];
         for (int i = 0; i < 8; i++) {
             extents[i] =
@@ -83,8 +83,8 @@ public class HfsPlusForkData {
     public byte[] write(byte[] dest, int destOffSet) {
         byte[] data = new byte[FORK_DATA_LENGTH];
         BigEndian.setInt64(data, 0, totalSize);
-        BigEndian.setInt32(data, 8, (int) clumpSize);
-        BigEndian.setInt32(data, 12, (int) totalBlock);
+        BigEndian.setInt32(data, 8, clumpSize);
+        BigEndian.setInt32(data, 12, totalBlock);
         for (int i = 0; i < extents.length; i++) {
             extents[i].write(data, EXTENT_OFFSET + (i * ExtentDescriptor.EXTENT_DESCRIPTOR_LENGTH));
         }
@@ -107,11 +107,11 @@ public class HfsPlusForkData {
         return totalSize;
     }
 
-    public long getClumpSize() {
+    public int getClumpSize() {
         return clumpSize;
     }
 
-    public long getTotalBlocks() {
+    public int getTotalBlocks() {
         return totalBlock;
     }
